+++ runtime/src/utils/mod.rs
@@ -5,7 +5,6 @@ use crate::Runner;
 use crate::{Metrics, Spawner};
 #[cfg(test)]
 use futures::stream::{FuturesUnordered, StreamExt};
-use futures::{channel::oneshot, future::Shared, FutureExt};
 use rayon::{ThreadPool as RThreadPool, ThreadPoolBuildError, ThreadPoolBuilder};
 use std::{
     any::Any,
@@ -16,6 +15,7 @@ use std::{
 };
 
 pub mod buffer;
+pub mod signal;
 
 mod handle;
 pub use handle::Handle;
@@ -50,221 +50,6 @@ fn extract_panic_message(err: &(dyn Any + Send)) -> String {
         s.clone()
     } else {
         format!("{err:?}")
-    }
-}
-
-/// A one-time broadcast that can be awaited by many tasks. It is often used for
-/// coordinating shutdown across many tasks.
-///
-/// Each Signal tracks its lifecycle to enable proper shutdown coordination.
-/// To minimize overhead, it is recommended to wait on a reference to it
-/// (i.e. `&mut signal`) in loops rather than creating multiple `Signal`s.
-///
-/// # Example
-///
-/// ## Basic Usage
-///
-/// ```rust
-/// use commonware_runtime::{Spawner, Runner, Signaler, deterministic};
-///
-/// let executor = deterministic::Runner::default();
-/// executor.start(|context| async move {
-///     // Setup signaler and get future
-///     let (signaler, signal) = Signaler::new();
-///
-///     // Signal shutdown
-///     signaler.signal(2);
-///
-///     // Wait for shutdown in task
-///     let sig = signal.await.unwrap();
-///     println!("Received signal: {}", sig);
-/// });
-/// ```
-///
-/// ## Advanced Usage
-///
-/// While `Futures::Shared` is efficient, there is still meaningful overhead
-/// to cloning it (i.e. in each iteration of a loop). To avoid
-/// a performance regression from introducing `Signaler`, it is recommended
-/// to wait on a reference to `Signal` (i.e. `&mut signal`).
-///
-/// ```rust
-/// use commonware_macros::select;
-/// use commonware_runtime::{Clock, Spawner, Runner, Signaler, deterministic, Metrics};
-/// use futures::channel::oneshot;
-/// use std::time::Duration;
-///
-/// let executor = deterministic::Runner::default();
-/// executor.start(|context| async move {
-///     // Setup signaler and get future
-///     let (signaler, mut signal) = Signaler::new();
-///
-///     // Loop on the signal until resolved
-///     let (tx, rx) = oneshot::channel();
-///     context.with_label("waiter").spawn(|context| async move {
-///         loop {
-///             // Wait for signal or sleep
-///             select! {
-///                  sig = &mut signal => {
-///                      println!("Received signal: {}", sig.unwrap());
-///                      break;
-///                  },
-///                  _ = context.sleep(Duration::from_secs(1)) => {},
-///             };
-///         }
-///         let _ = tx.send(());
-///     });
-///
-///     // Send signal
-///     signaler.signal(9);
-///
-///     // Wait for task
-///     rx.await.expect("shutdown signaled");
-/// });
-/// ```
-#[derive(Clone)]
-pub enum Signal {
-    /// A live signal that will resolve when the signaler signals completion.
-    Live(LiveSignal),
-    /// A pre-resolved signal with a known value.
-    Resolved { value: i32 },
-}
-
-/// A live signal with completion tracking.
-#[derive(Clone)]
-pub struct LiveSignal {
-    inner: Shared<oneshot::Receiver<i32>>,
-    _guard: Arc<SignalGuard>,
-}
-
-struct SignalGuard {
-    tx: Option<oneshot::Sender<()>>,
-}
-
-impl Drop for SignalGuard {
-    fn drop(&mut self) {
-        if let Some(tx) = self.tx.take() {
-            let _ = tx.send(());
-        }
-    }
-}
-
-impl Future for Signal {
-    type Output = Result<i32, oneshot::Canceled>;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        match &mut *self {
-            Signal::Live(live) => Pin::new(&mut live.inner).poll(cx),
-            Signal::Resolved { value } => Poll::Ready(Ok(*value)),
-        }
-    }
-}
-
-/// Coordinates a one-time signal across many tasks.
-pub struct Signaler {
-    tx: oneshot::Sender<i32>,
-    completion_rx: oneshot::Receiver<()>,
-}
-
-impl Signaler {
-    /// Create a new `Signaler`.
-    ///
-    /// Returns a `Signaler` and a `Signal` that will resolve when `signal` is called.
-    pub fn new() -> (Self, Signal) {
-        let (tx, rx) = oneshot::channel();
-        let (completion_tx, completion_rx) = oneshot::channel();
-
-        let signaler = Self { tx, completion_rx };
-        let signal = Signal::Live(LiveSignal {
-            inner: rx.shared(),
-            _guard: Arc::new(SignalGuard {
-                tx: Some(completion_tx),
-            }),
-        });
-
-        (signaler, signal)
-    }
-
-    /// Resolve all `Signal`s associated with this `Signaler`.
-    pub fn signal(self, value: i32) -> oneshot::Receiver<()> {
-        let _ = self.tx.send(value);
-        self.completion_rx
-    }
-}
-
-/// Manages shutdown state for a runtime executor.
-pub enum ShutdownState {
-    /// The runtime is running and stop has not been called yet.
-    Running { signaler: Signaler, signal: Signal },
-    /// Stop has been called but completion is still pending.
-    Stopping {
-        stop_value: i32,
-        completion: Shared<oneshot::Receiver<()>>,
-    },
-    /// Stop has completed.
-    Stopped { stop_value: i32 },
-}
-
-impl ShutdownState {
-    /// Create a new shutdown state in running mode.
-    pub fn new() -> Self {
-        let (signaler, signal) = Signaler::new();
-        Self::Running { signaler, signal }
-    }
-
-    /// Get the signal for runtime users to await.
-    pub fn stopped(&self) -> Signal {
-        match self {
-            Self::Running { signal, .. } => signal.clone(),
-            Self::Stopping { stop_value, .. } | Self::Stopped { stop_value } => {
-                Signal::Resolved { value: *stop_value }
-            }
-        }
-    }
-
-    /// Initiate shutdown returning a completion future.
-    /// Returns `None` if stop has already completed.
-    pub fn stop(&mut self, value: i32) -> Option<Shared<oneshot::Receiver<()>>> {
-        match std::mem::replace(self, Self::Stopped { stop_value: value }) {
-            Self::Running { signaler, .. } => {
-                let completion_rx = signaler.signal(value);
-                let shared_completion = completion_rx.shared();
-                *self = Self::Stopping {
-                    stop_value: value,
-                    completion: shared_completion.clone(),
-                };
-                Some(shared_completion)
-            }
-            Self::Stopping {
-                stop_value,
-                completion,
-            } => {
-                *self = Self::Stopping {
-                    stop_value,
-                    completion: completion.clone(),
-                };
-                Some(completion)
-            }
-            Self::Stopped { stop_value } => {
-                *self = Self::Stopped { stop_value };
-                None
-            }
-        }
-    }
-
-    /// Mark shutdown as finished. This should be called after the completion future resolves.
-    pub fn finished(&mut self) {
-        if let Self::Stopping { stop_value, .. } = self {
-            *self = Self::Stopped {
-                stop_value: *stop_value,
-            };
-        }
-    }
-}
-
-impl Default for ShutdownState {
-    fn default() -> Self {
-        Self::new()
     }
 }
